+++ rootfs/app/docker-entrypoint.sh
@@ -10,4 +10,17 @@ then
   cd /appV2
 fi
 
-java -jar senzing-api-server.jar "$@"
+# Remove "java -jar senzing-api-server.jar" from input parameters.
+
+PARAMETERS=$(
+    echo "$@" \
+    | sed \
+        -e s/java// \
+        -e s/-jar// \
+        -e s/senzing-poc-server.jar//
+)
+TOKENIZED_PARAMETERS=(${PARAMETERS})
+
+# Run Java command.
+
+java -jar senzing-poc-server.jar "${TOKENIZED_PARAMETERS[@]}"