+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed in 3.4.2
 
-- Fixed Docker command-line invocation behavior
+- In `Dockerfile`, removed `CMD` to simplify input parameters
 
 ## [3.4.1] - 2022-09-28